requirements.txt
requests
beautifulsoup4
pdfminer.six
openai
